% Generated by roxygen2 (4.0.0): do not edit by hand
\name{ggSurv}
\alias{ggSurv}
\title{ggplot KM survival curves, stratified by variable of choice}
\usage{
ggSurv(EventTime, EventStatus, nam, dat)
}
\arguments{
  \item{surv}{EventTime survival times EventStatus
  censoring indicator (1 implies event observed, 0
  otherwise) nam string literal of variable in dat
  indicating variable by which KM curves should be
  stratified dat dataframe with above 3 mentioned
  covariates}
}
\description{
ggplot KM survival curves, stratified by variable of choice
}
\examples{
g <- ggSurv(EventTime,EventStatus,'plink_rs2716212_G',dat)
ggsave(g,file='~/test_gg.jpg',height=6,width=8)
}
\keyword{curve}
\keyword{kaplan-meier}
\keyword{km}
\keyword{survival}

